Move cpuid facility to separate package
